+++ src/earley/mod.rs
@@ -271,260 +271,6 @@ mod tests {
     use crate::Grammar;
     use quickcheck::{Arbitrary, Gen, QuickCheck, TestResult};
 
-    #[test]
-    fn undefined_prod() {
-        let grammar: Grammar = "
-        <start> ::= <a> | <b>
-        <a> ::= 'a'
-        "
-        .parse()
-        .unwrap();
-
-        let input = "a";
-
-        let parses = parse(&grammar, input);
-        assert_eq!(parses.count(), 1);
-    }
-
-    #[test]
-    fn dna_left_recursive() {
-        let grammar: Grammar = "<dna> ::= <base> | <dna> <base>
-        <base> ::= \"A\" | \"C\" | \"G\" | \"T\""
-            .parse()
-            .unwrap();
-
-        let input = "GATTACA";
-
-        let parses = parse(&grammar, input);
-        assert_eq!(parses.count(), 1);
-    }
-
-    #[test]
-    fn dna_right_recursive() {
-        let grammar: Grammar = "<dna> ::= <base> | <base> <dna>
-        <base> ::= \"A\" | \"C\" | \"G\" | \"T\""
-            .parse()
-            .unwrap();
-
-        let input = "GATTACA";
-
-        let parses = parse(&grammar, input);
-        assert_eq!(parses.count(), 1);
-    }
-
-    #[test]
-    fn ambiguous() {
-        let grammar: Grammar = "<start> ::= <a> | <b>
-        <a> ::= \"END\"
-        <b> ::= \"END\""
-            .parse()
-            .unwrap();
-
-        let input = "END";
-
-        let parses = parse(&grammar, input);
-        assert_eq!(parses.count(), 2);
-    }
-
-    #[test]
-    fn recursive_nested_infinite() {
-        let grammar: Grammar = "
-            <a> ::= <b> | 'z'
-            <b> ::= <a> 
-        "
-        .parse()
-        .unwrap();
-
-        let input = "z";
-
-        // there are infinite parses to this, so take the first 100 and call it good
-        // TODO: why does stack explode so quickly?
-        let parse_count = 100;
-        let parses = parse(&grammar, input).take(parse_count);
-        assert_eq!(parses.count(), parse_count);
-    }
-
-    #[test]
-    fn empty_right_recursive() {
-        let grammar: Grammar = "<a> ::= '' | 'a' <a>".parse().unwrap();
-
-        let input = "aaaaaaaaaa";
-
-        let parses = parse(&grammar, input);
-        assert_eq!(parses.count(), 1);
-    }
-
-    #[test]
-    fn empty_left_recursive() {
-        let grammar: Grammar = "<a> ::= '' | <a> 'a'".parse().unwrap();
-
-        let input = "aaaaaaaaaa";
-
-        let parses = parse(&grammar, input);
-        assert_eq!(parses.count(), 1);
-    }
-
-    #[test]
-    fn complete_empty() {
-        let grammar: Grammar = "<start> ::= \"hi\" <empty>
-        <empty> ::= \"\""
-            .parse()
-            .unwrap();
-
-        let input = "hi";
-
-        let parses = parse(&grammar, input);
-        assert_eq!(parses.count(), 1);
-    }
-
-    #[test]
-    fn empty() {
-        let grammar: Grammar = "<start> ::= \"\"".parse().unwrap();
-
-        let input = "";
-
-        let parses = parse(&grammar, input);
-        assert_eq!(parses.count(), 1);
-    }
-
-    #[test]
-    fn nested_empty_post() {
-        let grammar: Grammar = "
-        <start> ::= <a> <empty>
-        <a> ::= 'a' <empty>
-        <empty> ::= ''"
-            .parse()
-            .unwrap();
-
-        let input = "a";
-
-        let parses = parse(&grammar, input);
-        assert_eq!(parses.count(), 1);
-    }
-
-    #[test]
-    fn nested_empty_pre() {
-        let grammar: Grammar = "
-        <start> ::= <empty> <a>
-        <a> ::= <empty> 'a'
-        <empty> ::= ''"
-            .parse()
-            .unwrap();
-
-        let input = "a";
-
-        let parses = parse(&grammar, input);
-        assert_eq!(parses.count(), 1);
-    }
-
-    #[test]
-    fn nested_empty_pre_and_post() {
-        let grammar: Grammar = "
-        <start> ::= <empty> <a> <empty>
-        <a> ::= <empty> 'a' <empty>
-        <empty> ::= ''"
-            .parse()
-            .unwrap();
-
-        let input = "a";
-
-        let parses = parse(&grammar, input);
-        assert_eq!(parses.count(), 1);
-    }
-
-    #[test]
-    fn empty_inline() {
-        let grammar: Grammar = "
-        <start> ::= <a> '' <a>
-        <a> ::= 'a'"
-            .parse()
-            .unwrap();
-
-        let input = "aa";
-
-        let parses = parse(&grammar, input);
-        assert_eq!(parses.count(), 1);
-    }
-
-    #[test]
-    fn empty_ambiguous() {
-        let grammar: Grammar = "
-        <start> ::= <a> | <b>
-        <a> ::= ''
-        <b> ::= ''"
-            .parse()
-            .unwrap();
-
-        let input = "";
-
-        let parses = parse(&grammar, input);
-        assert_eq!(parses.count(), 2);
-    }
-
-    #[test]
-    fn empty_first_nested() {
-        // this structure exposes improper "nullable" production detection
-        let grammar: Grammar = "
-        <a> ::= '' | '' <b> <c>
-        <b> ::= <c>
-        <c> ::= <a>
-        "
-        .parse()
-        .unwrap();
-
-        let input = "";
-
-        let parse_count = 100;
-        let parses = parse(&grammar, input).take(parse_count);
-        assert_eq!(parses.count(), parse_count);
-    }
-
-    #[test]
-    fn optional_whitespace() {
-        let grammar: Grammar = "
-        <balanced> ::= <left> <balanced> <right>
-                     | ''
-        
-        <left> ::= <opt-ws> '(' <opt-ws>
-        <right> ::= <opt-ws> ')' <opt-ws>
-        
-        <opt-ws> ::= '' | <ws>
-        <ws> ::= ' ' | ' ' <ws>
-        "
-        .parse()
-        .unwrap();
-
-        let input = "()";
-
-        assert!(
-            grammar.parse_input(input).next().is_some(),
-            "can't parse: {input}"
-        );
-    }
-
-    #[test]
-    fn qualified_whitespace() {
-        let grammar: Grammar = "
-        <terms> ::= <terms> <ws> <term>
-                  | <term>
-        <term> ::= <qualified>
-                 | 'unqualified'
-        <qualified> ::= 'QUALIFIER:' <qual-term>
-        <qual-term> ::= <qual-term> <ws>
-                      | 'qualified'
-        <ws> ::= ' ' | ' ' <ws>
-        "
-        .parse()
-        .unwrap();
-
-        let input = "QUALIFIER:qualified unqualified";
-
-        assert!(
-            grammar.parse_input(input).next().is_some(),
-            "can't parse: {input}"
-        );
-    }
-
     #[derive(Debug, Clone)]
     struct NestedEmptyGrammar(Grammar);
     impl Arbitrary for NestedEmptyGrammar {
@@ -569,91 +315,5 @@ mod tests {
         QuickCheck::new()
             .tests(1000)
             .quickcheck(prop_empty_rules_allow_parse as fn(NestedEmptyGrammar) -> TestResult)
-    }
-
-    // (source: <https://loup-vaillant.fr/tutorials/earley-parsing/empty-rules>)
-    #[test]
-    fn empty_noop_infinite() {
-        let grammar: Grammar = "
-        <a> ::= '' | <b>
-        <b> ::= <a>"
-            .parse()
-            .unwrap();
-
-        let input = "";
-
-        let parse_count = 100;
-        let parses = parse(&grammar, input).take(parse_count);
-        assert_eq!(parses.count(), parse_count);
-    }
-
-    // (source: <https://loup-vaillant.fr/tutorials/earley-parsing/recogniser>)
-    // Sum     -> Sum     [+-] Product
-    // Sum     -> Product
-    // Product -> Product [*/] Factor
-    // Product -> Factor
-    // Factor  -> '(' Sum ')'
-    // Factor  -> Number
-    // Number  -> [0-9] Number
-    // Number  -> [0-9]
-    #[test]
-    fn math() {
-        let grammar: Grammar = "<sum> ::= <sum> <add> <product>
-            <sum> ::= <product>
-            <product> ::= <product> <mult> <factor>
-            <product> ::= <factor>
-            <add> ::= \"+\" | \"-\"
-            <mult> ::= \"*\" | \"/\"
-            <factor> ::= \"(\" <sum> \")\"
-            <factor> ::= <number>
-            <number> ::= <digit> <number>
-            <number> ::= <digit>
-            <digit> ::= \"0\" | \"1\" | \"2\" | \"3\" | \"4\" | \"5\" | \"6\" | \"7\" | \"8\" | \"9\"
-        ".parse().unwrap();
-
-        let input = "1+(2*3-4)";
-
-        let parses: Vec<_> = parse(&grammar, input).collect();
-
-        let expected_parse_tree = "
-<sum> ::= <sum> <add> <product>
-├── <sum> ::= <product>
-│   └── <product> ::= <factor>
-│       └── <factor> ::= <number>
-│           └── <number> ::= <digit>
-│               └── <digit> ::= \"1\"
-│                   └── \"1\"
-├── <add> ::= \"+\"
-│   └── \"+\"
-└── <product> ::= <factor>
-    └── <factor> ::= \"(\" <sum> \")\"
-        ├── \"(\"
-        ├── <sum> ::= <sum> <add> <product>
-        │   ├── <sum> ::= <product>
-        │   │   └── <product> ::= <product> <mult> <factor>
-        │   │       ├── <product> ::= <factor>
-        │   │       │   └── <factor> ::= <number>
-        │   │       │       └── <number> ::= <digit>
-        │   │       │           └── <digit> ::= \"2\"
-        │   │       │               └── \"2\"
-        │   │       ├── <mult> ::= \"*\"
-        │   │       │   └── \"*\"
-        │   │       └── <factor> ::= <number>
-        │   │           └── <number> ::= <digit>
-        │   │               └── <digit> ::= \"3\"
-        │   │                   └── \"3\"
-        │   ├── <add> ::= \"-\"
-        │   │   └── \"-\"
-        │   └── <product> ::= <factor>
-        │       └── <factor> ::= <number>
-        │           └── <number> ::= <digit>
-        │               └── <digit> ::= \"4\"
-        │                   └── \"4\"
-        └── \")\"\n"
-            .trim_start();
-
-        assert_eq!(parses.len(), 1);
-        let parse_tree = format!("{}", parses[0]);
-        assert_eq!(parse_tree, expected_parse_tree)
     }
 }
