+++ src/grammar.rs
@@ -680,208 +680,6 @@ mod tests {
     }
 
     #[test]
-    fn parse_dna() {
-        let grammar: Grammar = "<dna> ::= <base> | <base> <dna>
-        <base> ::= \"A\" | \"C\" | \"G\" | \"T\""
-            .parse()
-            .unwrap();
-
-        let input = "GATTACA";
-
-        let mut parses = grammar.parse_input(input);
-        assert!(matches!(parses.next(), Some(_)));
-    }
-
-    #[test]
-    fn shared_nonterminal_failure() {
-        let grammar = "
-        <start> ::= <shortfail> | <longsuccess>
-        <shortfail> ::= <char> 'never'
-        <char> ::= 'a'
-        <longsuccess> ::= <long2>
-        <long2> ::= <long3>
-        <long3> ::= <long4>
-        <long4> ::= <char>
-        ";
-
-        let grammar = grammar.parse::<Grammar>().unwrap();
-
-        let input = "a";
-
-        assert!(
-            grammar.parse_input(input).next().is_some(),
-            "minimal issue failed to parse: {input}"
-        );
-    }
-
-    #[test]
-    fn swap_left_right_recursion() {
-        let input = "aa a";
-
-        let left_recursive: &str = "
-        <conjunction> ::= <conjunction> <ws> <predicate> | <predicate>
-        <predicate> ::= <string_null_one> | <special-string> '.'
-        <string_null_one> ::= <string_null_two>
-        <string_null_two> ::= <string_null_three>
-        <string_null_three> ::= <string>
-        <string> ::= <char_null> | <string> <char_null>
-        <special-string> ::= <special-char> | <special-string> <special-char>
-        <char_null> ::= <char>
-        <char> ::= 'a'
-        <special-char> ::= <char_null> | <whitespace>
-        <whitespace> ::= ' '
-        <ws> ::= ' ' | ' ' <ws>
-        ";
-        assert!(left_recursive
-            .parse::<Grammar>()
-            .unwrap()
-            .parse_input(input)
-            .next()
-            .is_some());
-
-        let right_recursive = left_recursive.replace(
-            // rewrite production from left- to right- recursive
-            "<string> ::= <char_null> | <string> <char_null>",
-            "<string> ::= <char_null> | <char_null> <string>",
-        );
-        assert!(
-            right_recursive
-                .parse::<Grammar>()
-                .unwrap()
-                .parse_input(input)
-                .next()
-                .is_some(),
-            "right-recursive grammar failed to parse: {input}"
-        );
-    }
-
-    #[test]
-    fn shared_nullable_nonterminal() {
-        let fails: &str = "
-        <disjunction> ::= <predicate> | <disjunction> <or> <predicate>
-        <predicate> ::= <char_null_one> | <special-string> '.'
-
-        <char_null_one> ::= <char_null_two>
-        <char_null_two> ::= <char_null_three>
-        <char_null_three> ::= <char>
-
-        <or> ::= <ws> 'or' <ws>
-        <ws> ::= <whitespace> | ' ' <ws>
-        <whitespace> ::= ' '
-
-        <special-string> ::= <special-char> | <special-char> <special-string>
-        <special-char> ::= <char> | <whitespace>
-        <char> ::= 'a'
-        ";
-
-        let input = "a or a";
-
-        let passes_1 = fails.replace(
-            // skip nullable production <char_null_two>
-            "<char_null_one> ::= <char_null_two>",
-            "<char_null_one> ::= <char_null_three>",
-        );
-        assert!(passes_1
-            .parse::<Grammar>()
-            .unwrap()
-            .parse_input(input)
-            .next()
-            .is_some());
-
-        let passes_2 = fails.replace(
-            // replace <whitespace> with its terminal ' '
-            "<ws> ::= <whitespace> | ' ' <ws>",
-            "<ws> ::= ' ' | ' ' <ws>",
-        );
-        assert!(passes_2
-            .parse::<Grammar>()
-            .unwrap()
-            .parse_input(input)
-            .next()
-            .is_some());
-
-        let passes_3 = fails.replace(
-            // again, replace <whitespace> with its terminal ' '
-            "<special-char> ::= <char> | <whitespace>",
-            "<special-char> ::= <char> | ' '",
-        );
-        assert!(passes_3
-            .parse::<Grammar>()
-            .unwrap()
-            .parse_input(input)
-            .next()
-            .is_some());
-
-        assert!(
-            fails
-                .parse::<Grammar>()
-                .unwrap()
-                .parse_input(input)
-                .next()
-                .is_some(),
-            "all grammars except last one parsed: {input}"
-        );
-    }
-
-    #[test]
-    fn branching_and_overlapping_parse_trees() {
-        let bnf = "
-        <and> ::= <and> ' AND ' <terminal>
-                | <and> ' ' <terminal>
-                | <terminal>
-        <terminal> ::= 'AND'
-        ";
-        let input = "AND AND AND AND AND";
-
-        // 1. 'AND' <and> 'AND' <and> 'AND'
-        // 2. 'AND' <and> 'AND' 'AND' 'AND'
-        // 3. 'AND' 'AND' <and> 'AND' 'AND'
-        // 4. 'AND' 'AND' 'AND' <and> 'AND'
-        // 5. 'AND' 'AND' 'AND' 'AND' 'AND'
-        assert_eq!(
-            bnf.parse::<Grammar>().unwrap().parse_input(input).count(),
-            5
-        );
-    }
-
-    #[test]
-    fn format_parse_tree() {
-        let grammar: Grammar = "<dna> ::= <base> | <base> <dna>
-        <base> ::= \"A\" | \"C\" | \"G\" | \"T\""
-            .parse()
-            .unwrap();
-
-        let input = "GATTACA";
-        let parsed = grammar.parse_input(input).next().unwrap();
-        let formatted = format!("{parsed}");
-        let expected = "
-<dna> ::= <base> <dna>
-├── <base> ::= \"G\"
-│   └── \"G\"
-└── <dna> ::= <base> <dna>
-    ├── <base> ::= \"A\"
-    │   └── \"A\"
-    └── <dna> ::= <base> <dna>
-        ├── <base> ::= \"T\"
-        │   └── \"T\"
-        └── <dna> ::= <base> <dna>
-            ├── <base> ::= \"T\"
-            │   └── \"T\"
-            └── <dna> ::= <base> <dna>
-                ├── <base> ::= \"A\"
-                │   └── \"A\"
-                └── <dna> ::= <base> <dna>
-                    ├── <base> ::= \"C\"
-                    │   └── \"C\"
-                    └── <dna> ::= <base>
-                        └── <base> ::= \"A\"
-                            └── \"A\"\n"
-            .trim_start();
-
-        assert_eq!(formatted, expected);
-    }
-
-    #[test]
     fn format_grammar() {
         let grammar: Grammar = "<dna> ::= <base> | <base> <dna>
         <base> ::= \"A\" | \"C\" | \"G\" | \"T\""
@@ -926,101 +724,5 @@ mod tests {
         *rhs_iterated = ParseTreeNode::Terminal("Z");
 
         assert_eq!(parse_tree.rhs[0], ParseTreeNode::Terminal("Z"));
-    }
-
-    #[test]
-    fn mermaid_dna_parse_tree() {
-        let grammar: Grammar = "<dna> ::= <base> | <base> <dna>
-        <base> ::= \"A\" | \"C\" | \"G\" | \"T\""
-            .parse()
-            .unwrap();
-
-        let input = "GATTACA";
-        let parsed = grammar.parse_input(input).next().unwrap();
-        let mermaid = parsed.mermaid().to_string();
-        let expected = "
-flowchart TD
-0[\"dna\"] --> 1[\"base\"]
-1[\"base\"] --> 2[\"G\"]
-0[\"dna\"] --> 3[\"dna\"]
-3[\"dna\"] --> 4[\"base\"]
-4[\"base\"] --> 5[\"A\"]
-3[\"dna\"] --> 6[\"dna\"]
-6[\"dna\"] --> 7[\"base\"]
-7[\"base\"] --> 8[\"T\"]
-6[\"dna\"] --> 9[\"dna\"]
-9[\"dna\"] --> 10[\"base\"]
-10[\"base\"] --> 11[\"T\"]
-9[\"dna\"] --> 12[\"dna\"]
-12[\"dna\"] --> 13[\"base\"]
-13[\"base\"] --> 14[\"A\"]
-12[\"dna\"] --> 15[\"dna\"]
-15[\"dna\"] --> 16[\"base\"]
-16[\"base\"] --> 17[\"C\"]
-15[\"dna\"] --> 18[\"dna\"]
-18[\"dna\"] --> 19[\"base\"]
-19[\"base\"] --> 20[\"A\"]\n"
-            .trim_start();
-
-        assert_eq!(mermaid, expected);
-    }
-
-    #[test]
-    fn mermaid_math_parse_tree() {
-        let grammar: Grammar = "<sum> ::= <sum> <add> <product>
-            <sum> ::= <product>
-            <product> ::= <product> <mult> <factor>
-            <product> ::= <factor>
-            <add> ::= \"+\" | \"-\"
-            <mult> ::= \"*\" | \"/\"
-            <factor> ::= \"(\" <sum> \")\"
-            <factor> ::= <number>
-            <number> ::= <digit> <number>
-            <number> ::= <digit>
-            <digit> ::= \"0\" | \"1\" | \"2\" | \"3\" | \"4\" | \"5\" | \"6\" | \"7\" | \"8\" | \"9\"
-        ".parse().unwrap();
-
-        let input = "1+(2*3-4)";
-
-        let parsed = grammar.parse_input(input).next().unwrap();
-        let mermaid = parsed.mermaid().to_string();
-        let expected = "
-flowchart TD
-0[\"sum\"] --> 1[\"sum\"]
-1[\"sum\"] --> 2[\"product\"]
-2[\"product\"] --> 3[\"factor\"]
-3[\"factor\"] --> 4[\"number\"]
-4[\"number\"] --> 5[\"digit\"]
-5[\"digit\"] --> 6[\"1\"]
-0[\"sum\"] --> 7[\"add\"]
-7[\"add\"] --> 8[\"+\"]
-0[\"sum\"] --> 9[\"product\"]
-9[\"product\"] --> 10[\"factor\"]
-10[\"factor\"] --> 11[\"(\"]
-10[\"factor\"] --> 12[\"sum\"]
-12[\"sum\"] --> 13[\"sum\"]
-13[\"sum\"] --> 14[\"product\"]
-14[\"product\"] --> 15[\"product\"]
-15[\"product\"] --> 16[\"factor\"]
-16[\"factor\"] --> 17[\"number\"]
-17[\"number\"] --> 18[\"digit\"]
-18[\"digit\"] --> 19[\"2\"]
-14[\"product\"] --> 20[\"mult\"]
-20[\"mult\"] --> 21[\"*\"]
-14[\"product\"] --> 22[\"factor\"]
-22[\"factor\"] --> 23[\"number\"]
-23[\"number\"] --> 24[\"digit\"]
-24[\"digit\"] --> 25[\"3\"]
-12[\"sum\"] --> 26[\"add\"]
-26[\"add\"] --> 27[\"-\"]
-12[\"sum\"] --> 28[\"product\"]
-28[\"product\"] --> 29[\"factor\"]
-29[\"factor\"] --> 30[\"number\"]
-30[\"number\"] --> 31[\"digit\"]
-31[\"digit\"] --> 32[\"4\"]
-10[\"factor\"] --> 33[\")\"]\n"
-            .trim_start();
-
-        assert_eq!(mermaid, expected);
     }
 }
