+++ tests/snapshots/parse_input__branching_and_overlapping_parse_trees.snap.new
@@ -1,0 +1,85 @@
+---
+source: tests/parse_input.rs
+assertion_line: 477
+expression: "parses.join(\"\\n\")"
+---
+<and> ::= <and> " AND " <terminal>
+├── <and> ::= <and> " AND " <terminal>
+│   ├── <and> ::= <terminal>
+│   │   └── <terminal> ::= "AND"
+│   │       └── "AND"
+│   ├── " AND "
+│   └── <terminal> ::= "AND"
+│       └── "AND"
+├── " AND "
+└── <terminal> ::= "AND"
+    └── "AND"
+
+<and> ::= <and> " AND " <terminal>
+├── <and> ::= <and> " " <terminal>
+│   ├── <and> ::= <and> " " <terminal>
+│   │   ├── <and> ::= <terminal>
+│   │   │   └── <terminal> ::= "AND"
+│   │   │       └── "AND"
+│   │   ├── " "
+│   │   └── <terminal> ::= "AND"
+│   │       └── "AND"
+│   ├── " "
+│   └── <terminal> ::= "AND"
+│       └── "AND"
+├── " AND "
+└── <terminal> ::= "AND"
+    └── "AND"
+
+<and> ::= <and> " " <terminal>
+├── <and> ::= <and> " " <terminal>
+│   ├── <and> ::= <and> " AND " <terminal>
+│   │   ├── <and> ::= <terminal>
+│   │   │   └── <terminal> ::= "AND"
+│   │   │       └── "AND"
+│   │   ├── " AND "
+│   │   └── <terminal> ::= "AND"
+│   │       └── "AND"
+│   ├── " "
+│   └── <terminal> ::= "AND"
+│       └── "AND"
+├── " "
+└── <terminal> ::= "AND"
+    └── "AND"
+
+<and> ::= <and> " " <terminal>
+├── <and> ::= <and> " AND " <terminal>
+│   ├── <and> ::= <and> " " <terminal>
+│   │   ├── <and> ::= <terminal>
+│   │   │   └── <terminal> ::= "AND"
+│   │   │       └── "AND"
+│   │   ├── " "
+│   │   └── <terminal> ::= "AND"
+│   │       └── "AND"
+│   ├── " AND "
+│   └── <terminal> ::= "AND"
+│       └── "AND"
+├── " "
+└── <terminal> ::= "AND"
+    └── "AND"
+
+<and> ::= <and> " " <terminal>
+├── <and> ::= <and> " " <terminal>
+│   ├── <and> ::= <and> " " <terminal>
+│   │   ├── <and> ::= <and> " " <terminal>
+│   │   │   ├── <and> ::= <terminal>
+│   │   │   │   └── <terminal> ::= "AND"
+│   │   │   │       └── "AND"
+│   │   │   ├── " "
+│   │   │   └── <terminal> ::= "AND"
+│   │   │       └── "AND"
+│   │   ├── " "
+│   │   └── <terminal> ::= "AND"
+│   │       └── "AND"
+│   ├── " "
+│   └── <terminal> ::= "AND"
+│       └── "AND"
+├── " "
+└── <terminal> ::= "AND"
+    └── "AND"
+
