+++ tests/snapshots/parse_input__empty_first_nested.snap.new
@@ -1,0 +1,52 @@
+---
+source: tests/parse_input.rs
+assertion_line: 218
+expression: "parses[0..=3].join(\"\\n\")"
+---
+<a> ::= ""
+└── ""
+
+<a> ::= "" <b> <c>
+├── ""
+├── <b> ::= <c>
+│   └── <c> ::= <a>
+│       └── <a> ::= ""
+│           └── ""
+└── <c> ::= <a>
+    └── <a> ::= ""
+        └── ""
+
+<a> ::= "" <b> <c>
+├── ""
+├── <b> ::= <c>
+│   └── <c> ::= <a>
+│       └── <a> ::= ""
+│           └── ""
+└── <c> ::= <a>
+    └── <a> ::= "" <b> <c>
+        ├── ""
+        ├── <b> ::= <c>
+        │   └── <c> ::= <a>
+        │       └── <a> ::= ""
+        │           └── ""
+        └── <c> ::= <a>
+            └── <a> ::= ""
+                └── ""
+
+<a> ::= "" <b> <c>
+├── ""
+├── <b> ::= <c>
+│   └── <c> ::= <a>
+│       └── <a> ::= "" <b> <c>
+│           ├── ""
+│           ├── <b> ::= <c>
+│           │   └── <c> ::= <a>
+│           │       └── <a> ::= ""
+│           │           └── ""
+│           └── <c> ::= <a>
+│               └── <a> ::= ""
+│                   └── ""
+└── <c> ::= <a>
+    └── <a> ::= ""
+        └── ""
+
